HTBHF-2521 move journey middleware to flow control directory
