+++ src/web/routes/application/check-answers/check-answers.js
@@ -1,6 +1,6 @@
 const { getCheckAnswers } = require('./get')
 const { CHECK_ANSWERS_URL } = require('../common/constants')
-const { handleRequestForPath } = require('../middleware')
+const { handleRequestForPath } = require('../flow-control')
 
 const registerCheckAnswersRoutes = (journey, config, app) => {
   app
