+++ src/web/routes/application/flow-control/index.js
@@ -1,3 +1,4 @@
 module.exports = {
-  ...require('./state-machine')
+  ...require('./state-machine'),
+  ...require('./middleware')
 }
