+++ src/web/routes/application/flow-control/middleware/session-details.js
